LPS-66761 Rename module
