+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-data-provider-instance/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Dynamic Data Mapping Data Provider Instance
+Bundle-SymbolicName: com.liferay.dynamic.data.mapping.data.provider.instance
+Bundle-Version: 1.0.0
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Dynamic Data Mapping
