--- modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-data-provider-osgi/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay Dynamic Data Mapping Data Provider OSGi
-Bundle-SymbolicName: com.liferay.dynamic.data.mapping.data.provider.osgi
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Dynamic Data Mapping
